refactoring messages according to MessageType enum
