+++ metadata-api/src/main/java/com/hashmap/haf/metadata/config/actors/message/metadata/MetadataMessage.java
@@ -1,14 +1,19 @@
-package com.hashmap.haf.metadata.config.actors.message;
+package com.hashmap.haf.metadata.config.actors.message.metadata;
 
+import com.hashmap.haf.metadata.config.actors.message.MessageType;
 import com.hashmap.haf.metadata.config.model.MetadataConfig;
 import lombok.Getter;
 
-abstract public class AbstractMetadataConfigMsg {
+public class MetadataMessage {
 
     @Getter
     private final MetadataConfig metadataConfig;
 
-    public AbstractMetadataConfigMsg(MetadataConfig metadataConfig) {
+    @Getter
+    private final MessageType messageType;
+
+    public MetadataMessage(MetadataConfig metadataConfig, MessageType messageType) {
         this.metadataConfig = metadataConfig;
+        this.messageType = messageType;
     }
 }
